Type casts from bools to numbers are now correct

+++ sc_data/src/protocol/convert.rs
@@ -269,7 +269,7 @@ pub fn type_cast(from: &RType, to: &RType) -> Vec<Op> {
     },
     "U" => return vec![],
     "i16" => match from.name.as_str() {
-      "bool" => Op::As(RType::new("u8")),
+      "bool" => Op::As(RType::new("i16")),
       "u8" | "i8" => into(),
       "i32" | "i64" => return try_into("i16"),
       "U" => return vec![],
@@ -282,7 +282,7 @@ pub fn type_cast(from: &RType, to: &RType) -> Vec<Op> {
       _ => panic!("cannot convert `{}` into `{}`", from, to),
     },
     "i32" => match from.name.as_str() {
-      "bool" => Op::As(RType::new("u8")),
+      "bool" => Op::As(RType::new("i32")),
       "f32" => Op::As(RType::new("i32")),
       "u8" | "i8" | "i16" => into(),
       "i64" => return try_into("i32"),
@@ -291,7 +291,7 @@ pub fn type_cast(from: &RType, to: &RType) -> Vec<Op> {
       _ => panic!("cannot convert `{}` into `{}`", from, to),
     },
     "i64" => match from.name.as_str() {
-      "bool" => Op::As(RType::new("u8")),
+      "bool" => Op::As(RType::new("i64")),
       "f32" => Op::As(RType::new("i32")),
       "u8" | "i8" | "i16" | "i32" => into(),
       _ => panic!("cannot convert `{}` into `{}`", from, to),
